# api/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Error:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  dto.Address:
    properties:
      city:
        type: string
      country:
        type: string
      street:
        type: string
    type: object
...
  dto.Supplier:
    properties:
      city:
        type: string
      country:
        type: string
      name:
        type: string
      phone_number:
        type: string
      street:
        type: string
    required:
    - name
    - phone_number
    type: object
host: aboba.com
info:
  contact: {}
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Swagger CRUD Home appliance store API
  version: "1.0"
paths:
...
  /api/v1/suppliers:
    get:
      consumes:
      - application/json
      description: That endpoint retrieve all registered supplier in system
      parameters:
      - description: limit get supplier
        in: query
        name: limit
        type: integer
      - description: offset get supplier
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Supplier'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Get all supplier
      tags:
      - suppliers
    post:
      consumes:
      - application/json
      description: 'Supplier created from JSON or XML, for create endpoint required:
        name, phone_number, country, city, street'
      parameters:
      - description: Supplier Data
        in: body
        name: supplier
        required: true
        schema:
          $ref: '#/definitions/dto.Supplier'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Create supplier
      tags:
      - suppliers
  /api/v1/suppliers/{id}:
    delete:
      consumes:
      - application/json
      description: That endpoint delete supplier data by id
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Delete supplier by ID
      tags:
      - suppliers
    get:
      consumes:
      - application/json
      description: That endpoint retrieve registered supplier in system by ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Supplier'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Get supplier by ID
      tags:
      - suppliers
    patch:
      consumes:
      - application/json
      description: That endpoint update supplier data (change address on supplier)
      parameters:
      - description: New address
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/dto.Address'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Update supplier by ID
      tags:
      - suppliers
swagger: "2.0"